requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
distro==1.9.0
fastapi==0.135.3
fonttools==4.62.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.14.0
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
openai==2.31.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
